+++ graphql/schema/index.js
@@ -1,0 +1,48 @@
+import { buildSchema } from "graphql";
+
+const schema = buildSchema(`
+type Event {
+    _id: ID!
+    title: String!
+    description:  String!
+    price: Float!
+    date: String!
+    creator: User!
+}
+
+input EventInput {
+    title: String!
+    description:  String!
+    price: Float!
+    date: String
+}
+
+type User {
+  _id: ID!,
+  email: String!,
+  password: String,
+  createdEvents:[Event!]
+}
+
+input UserInput {
+  email: String!,
+  password: String!
+}
+
+type RootQuery {
+    events: [Event!]!
+    users: [User]!
+}
+
+type RootMutation {
+    createEvent(eventInput: EventInput): Event
+    createUser(userInput: UserInput): User
+}
+
+schema {
+    query: RootQuery
+    mutation: RootMutation
+}
+`);
+
+export default schema;
